feat(header): обновил хедер и поправил верхний баннер
