+++ containers/header/header.module.scss
@@ -11,27 +11,67 @@
 
 .wrapper {
   width: 100%;
-  height: 80px;
+  height: 100px;
   @include mixins.layout;
   display: flex;
-  justify-content: space-between;
   align-items: center;
+  justify-content: space-between;
+
+  @include media.for-phone {
+    flex-direction: column-reverse;
+    align-items: center;
+    justify-content: center;
+  }
 }
 
 .logo {
-  color: red;
+  font-size: 12px;
+  font-weight: bold;
 
   img {
-    height: 40px;
+    height: 34px;
+  }
+
+  @include media.for-phone {
+    display: none;
   }
 }
+
+.center {
+  display: flex;
+  justify-content: space-between;
+  padding: 0 24px;
+  font-weight: bold;
+  font-size: 18px;
+  width: 70%;
+
+  @include media.for-laptop {
+    font-size: 14px;
+    padding: 0 16px;
+    flex-direction: column;
+    align-items: center;
+    text-align: center;
+
+    span:last-child {
+      margin-top: 4px;
+    }
+  }
+
+  @include media.for-phone {
+    font-size: 12px;
+  }
+}
+
 .phone {
-  font-weight: 700;
+  font-weight: bold;
   font-size: 36px;
   color: black;
-}
-.caption {
-  display: block;
-  font-weight: 400;
-  font-size: 18px;
+
+  @include media.for-phone {
+    margin-left: 0;
+    text-align: center;
+    width: 100%;
+    font-size: 32px;
+    margin-bottom: 8px;
+  }
 }
