+++ containers/header/header.tsx
@@ -6,13 +6,15 @@ export const Header: FC = () => {
   return (
     <div className={css.container}>
       <div className={css.wrapper}>
-        <div>
-          <a className={css.phone}>89073328743</a>
-          <span className={css.caption}>круглосуточно, без выходных</span>
-        </div>
         <div className={css.logo}>
           <img alt="logo" src="/eva-komi-logo.svg" />
+          <div>Сыктывкар и Республика Коми</div>
         </div>
+        <div className={css.center}>
+          <span>круглосуточно, без выходных</span>
+          <span>подача за 20 минут</span>
+        </div>
+        <a className={css.phone}>89073328743</a>
       </div>
       <ProgressBar />
     </div>
