+++ containers/page-home/banner/banner.module.scss
@@ -1,24 +1,42 @@
-@use 'styles/media';
+@use '/styles/media';
+@use '/styles/typography';
+@use '/styles/mixins';
 
 .banner {
   display: flex;
   flex-direction: column;
   align-items: center;
+  @include mixins.layout;
+
+  margin-top: 100px;
+
+  @include media.for-phone {
+    margin-top: 200px;
+  }
 }
+
+.carBanner {
+  max-width: 100%;
+}
+
+.content {
+  max-width: 600px;
+}
+
 .backgroundPhotoBanner {
   display: block;
   position: absolute;
-  left: 0%;
-  right: 0%;
+  left: 0;
+  right: 0;
   width: 100%;
   top: 250px;
   z-index: -1;
-}
-.carBanner {
-  width: 55%;
-  margin-top: 30px;
 
+  @include media.for-phone {
+    top: 300px;
+  }
 }
+
 .button {
   display: flex;
   flex-direction: column;
@@ -42,21 +60,12 @@
   box-shadow: none;
   transform: translateY(6px);
 }
-.textBanner {
-  font-size: 36px;
-  line-height: 42px;
+.title {
   text-align: center;
-  font-weight: 700;
-  padding-left: 10%;
-  padding-right: 10%;
-  max-width: 900px;
+  @include typography.h1;
 }
-.textBanner2 {
-  font-weight: 400;
-  font-size: 20px;
-  line-height: 23px;
+.text {
   text-align: center;
-  padding-left: 10%;
-  padding-right: 10%;
+  @include typography.body;
 }
