+++ styles/media.scss
@@ -1,7 +1,7 @@
 
 // медиа для ноутбука и маленького десктопного экрана
 @mixin for-laptop {
-  @media (min-width: 1024px) and (max-width: 1439px)  {
+  @media (max-width: 1439px)  {
     @content
   }
 }
